requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-chroma
python-dotenv
